make presets file for headers
