+++ src/index.js
@@ -8,7 +8,8 @@ export { default as ErrorText } from "./components/ErrorText";
 export { default as FileUpload } from "./components/FileUpload";
 export { default as GridCol } from "./components/GridCol";
 export { default as GridRow } from "./components/GridRow";
-export { default as Header, H1, H2, H3, H4, H5, H6 } from "./components/Header";
+export { default as Header } from "./components/Header";
+export { H1, H2, H3, H4, H5, H6 } from "./components/Header/presets";
 export { default as HintText } from "./components/HintText";
 export { default as Input } from "./components/Input";
 export { default as InputField } from "./components/InputField";
